Fixed wrong return type.

+++ src/StorageAttributor.h
@@ -14,7 +14,7 @@ class StorageAttributor
 {
 protected:
 	virtual const uint8_t GetPartitionsCount() { return 0; }
-	virtual const uint8_t GetPartitionSize(const uint8_t partition) { return 0; }
+	virtual const uint16_t GetPartitionSize(const uint8_t partition) { return 0; }
 
 public:
 	StorageAttributor() {}
